refactor(cli): extract compare.py pure formatters to _compare_format
Lifts the ten pure formatting helpers used by 'amx compare' rendering
and export (_fmt_dt, _fmt_duration, _fmt_or_dash, _confidence_style,
_band_prefix, _truncate, _fmt_int, _fmt_float, _md_escape, _md_table)
plus the _BAND_GLYPH constant into amx/cli_support/_compare_format.py.

The compare command re-exports each name so historical imports
(used by tests/test_compare.py and indirectly by web/routers/history.py)
keep working unchanged.

compare.py: 2509 -> 2426 LOC (-83, -3%). Modest but locks in a clean
foundation for follow-up extracts (CSV/JSON/Markdown exporters and
the per-column pivot renderer).
